# feat(electrum): optimize merkle proof validation with batching