--- a/crates/electrum/Cargo.toml
+++ b/crates/electrum/Cargo.toml
@@ -19,6 +19,7 @@ electrum-client = { version = "0.23.1", features = [ "proxy" ], default-features
 [dev-dependencies]
 bdk_testenv = { path = "../testenv" }
 bdk_chain = { path = "../chain" }
+criterion = { version = "0.2" }
 
 [features]
 default = ["use-rustls"]
@@ -29,3 +30,7 @@ use-openssl = ["electrum-client/use-openssl"]
 [[test]]
 name = "test_electrum"
 required-features = ["use-rustls"]
+
+[[bench]]
+name = "test_sync"
+harness = false