--- a/crates/electrum/src/bdk_electrum_client.rs
+++ b/crates/electrum/src/bdk_electrum_client.rs
@@ -339,6 +339,10 @@ impl<E: ElectrumApi> BdkElectrumClient<E> {
                     }
                 }
             }
+
+            if unused_spk_count >= stop_gap {
+                break;
+            }
         }
 
         Ok(last_active_index)