@ValuedMammal 2025-06-25
These are my criterion results after benching this PR. 👍 

```sh
sync_with_electrum      time:   [37.325 ms 38.220 ms 38.805 ms]                              
                        change: [-85.756% -85.556% -85.308%] (p = 0.00 < 0.05)
                        Performance has improved.
```